renamed TPTP AST datastructure from TPTPAST to TPTP
