+++ src/main/scala/leo/modules/agent/relevance_filter/BlackboardPreFilterSet.scala
@@ -1,7 +1,7 @@
 package leo.modules.agent.relevance_filter
 
 import leo.datastructures.blackboard._
-import leo.datastructures.TPTPAST.AnnotatedFormula
+import leo.datastructures.TPTP.AnnotatedFormula
 import leo.modules.relevance_filter.PreFilterSet
 
 /**
