+++ src/main/scala/leo/modules/modes/RunExternalProver.scala
@@ -2,7 +2,7 @@ package leo.modules.modes
 
 import leo.Configuration
 import leo.datastructures._
-import leo.datastructures.TPTPAST.AnnotatedFormula
+import leo.datastructures.TPTP.AnnotatedFormula
 import leo.modules.external.{Capabilities, ExternalProver}
 import leo.modules.output.SZS_Error
 import leo.modules.input.Input
