+++ src/main/scala/leo/modules/phase/InterleavableLoopPhase.scala
@@ -4,7 +4,7 @@ import leo.agents.{Agent, InterferingLoopAgent}
 import leo.datastructures.blackboard.{Blackboard, Delta}
 import leo.datastructures._
 import leo.datastructures.blackboard.scheduler.Scheduler
-import leo.datastructures.TPTPAST.AnnotatedFormula
+import leo.datastructures.TPTP.AnnotatedFormula
 import leo.modules.prover._
 import leo.modules.interleavingproc.{BlackboardState, SZSStatus, StateView}
 import leo.modules.control.Control
